pagonxt-onetradefinance-work/src/test/java/com/pagonxt/onetradefinance/work/api/backend/TaskOperationControllerTest.java
package com.pagonxt.onetradefinance.work.api.backend;

import com.fasterxml.jackson.databind.JsonNode;
import com.fasterxml.jackson.databind.ObjectMapper;
import com.pagonxt.onetradefinance.integrations.model.AuthenticatedRequest;
import com.pagonxt.onetradefinance.integrations.model.ControllerResponse;
import com.pagonxt.onetradefinance.work.config.ControllerTest;
import com.pagonxt.onetradefinance.work.service.TaskOperationService;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MvcResult;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@ControllerTest(TaskOperationController.class)
class TaskOperationControllerTest {
    @Autowired
    MockMvc mockMvc;

    @Autowired
    ObjectMapper mapper;

    @MockBean
    TaskOperationService taskOperationService;

    @Test
    void cancelCompleteInfoTask_ok_returnsOk() throws Exception {
        // Given
        // When and then
        mockMvc.perform(put("/backend/task-operation/complete-info/cancel/TSK-123")
                        .contentType(MediaType.APPLICATION_JSON)
                        .content("{}"))
                .andExpect(status().isOk()).andReturn();
    }

    @Test
    void cancelCompleteInfoTask_ok_returnsOKControllerResponse() throws Exception {
        // Given
        ControllerResponse expectedResult = ControllerResponse.success("cancelCompleteInfoTask", null);
        // When
        MvcResult mvcResult =
                mockMvc.perform(put("/backend/task-operation/complete-info/cancel/TSK-123")
                                .contentType(MediaType.APPLICATION_JSON)
                                .content("{}"))
                        .andReturn();

        // Then
        verify(taskOperationService, times(1)).cancelCompleteInfoTask("TSK-123", new AuthenticatedRequest());
        JsonNode resultJsonNode = mapper.readTree(mvcResult.getResponse().getContentAsString());
        JsonNode expectedJsonNode = mapper.valueToTree(expectedResult);
        assertThat(resultJsonNode).isEqualTo(expectedJsonNode);
    }
}
